Introducing ferretDB

+++ components/storage/test/migrations/1.7.0-1.test.js
@@ -33,6 +33,7 @@ describe('Migration - 1.7.x', function () {
   let webhooksCollection;
 
   before(async function () {
+    if (database.isFerret) this.skip();
     eventsCollection = await database.getCollection({ name: 'events' });
     usersCollection = await database.getCollection({ name: 'users' });
     streamsCollection = await database.getCollection({ name: 'streams' });
@@ -41,6 +42,7 @@ describe('Migration - 1.7.x', function () {
   });
 
   after(async function () {
+    if (database.isFerret) return;
     // erase all
     await eventsCollection.deleteMany({});
     await accessesCollection.deleteMany({});
